+++ aws_emr_launch/__init__.py
@@ -1,4 +1,5 @@
 import pkg_resources
 
-__version__ = pkg_resources.get_distribution("aws-emr-launch").version
-__package__ = f'aws-emr-launch-{__version__}'
+__product__ = 'aws-emr-launch'
+__version__ = pkg_resources.get_distribution(__product__).version
+__package__ = f'{__product__}-{__version__}'